Edited the QC to handle files with a single individual

+++ 1_QC_FileArgs.sh
@@ -6,6 +6,9 @@
 plink=$3
 export S=$4
 
+echo "##########################################################################"
+echo "Check individual call rate"
+echo "##########################################################################"
 ## Check for genotype call, heterozygosity
 $3 --file $1 --missing --het --cow --recode --out $1_$2
 
@@ -19,10 +22,21 @@ tail -n +2 $1_$2.imiss | awk '$6 >= 0.10 { print $1,$2 }' > IDsWithCallRateLessT
 ${S}/DrawHetMissPlotInd.R $1_$2
 echo "Created plot:$1_$2_imiss-vs-het.pdf"
 
-## Remove individuals with low genotype call rate and heterozygosity rate out of range +-6SD of heterozygosity
-$3 --file $1_$2 --remove IndividualsToExlcudeMissHet.txt --missing --recode --cow --out $1_$2_CleanInds
-mv IndividualsToExlcudeMissHet.txt IndividualsToExlcudeMissHet_$1_$2.txt
 
+## Remove individuals with low genotype call rate and heterozygosity rate out of range +-6SD of heterozygosity
+FILE=./IndividualsToExlcudeMissHet.txt
+if [ -f "$FILE" ]; then
+      $3 --file $1_$2 --remove IndividualsToExlcudeMissHet.txt --missing --recode --cow --out $1_$2_CleanInds
+      mv IndividualsToExlcudeMissHet.txt IndividualsToExlcudeMissHet_$1_$2.txt
+else 
+      mv $1_$2.ped $1_$2_CleanInds.ped
+      mv $1_$2.map $1_$2_CleanInds.map
+      mv $1_$2.lmiss  $1_$2_CleanInds.lmiss
+fi
+
+echo "##########################################################################"
+echo "Check for SNP call rate"
+echo "##########################################################################"
 ## Extract SNP with missing information on more than 10% of genotypes
 tail -n +2 $1_$2_CleanInds.lmiss | awk '$5 > 0.10 { print $1,$2 }' > MarkersWithCallRateLessThan0.90Plink_$1_$2.txt
